refactor(DTO): restructure DTO entities and service interactions
