+++ Client/Shared/Chat.razor
@@ -24,11 +24,7 @@
 	{
 		if (string.IsNullOrWhiteSpace(newMessageText)) return;
 
-		var newMessage = new ChatMessageDto
-		{
-			Author = CurrentUserName,
-			Text = newMessageText
-		};
+		var newMessage = new ChatMessageDto(CurrentUserName, newMessageText);
 
 		await SendMessage.InvokeAsync(newMessage);
