+++ Client/_Imports.razor
@@ -13,4 +13,5 @@
 @using Frontend.Client.Services
 @using Frontend.Shared
 @using Frontend.Shared.Models
+@using Frontend.Shared.Models.Dtos
 @using MudBlazor
