+++ Shared/Models/Dtos/ChatMessageDto.cs
@@ -1,0 +1,2 @@
+namespace Frontend.Shared.Models.Dtos;
+public record ChatMessageDto(string Author, string Text) { }
