+++ Shared/Models/Dtos/FriendRequestDto.cs
@@ -1,0 +1,3 @@
+namespace Frontend.Shared.Models.Dtos;
+
+public record FriendRequestDto(string ReceiverId, string SenderId, DateTime CreatedAt) { }
