+++ Shared/Models/Dtos/MoveDto.cs
@@ -1,0 +1,2 @@
+namespace Frontend.Shared.Models.Dtos;
+public record MoveDto(string From, string To, string? Promotion = null) { }
