+++ Shared/Models/Dtos/NotificationDto.cs
@@ -1,0 +1,3 @@
+namespace Frontend.Shared.Models.Dtos;
+
+public record NotificationDto(string Id, string Sender, string Message, bool IsRead, DateTime CreatedAt) { }
