+++ Shared/Models/Dtos/PlayerDto.cs
@@ -1,0 +1,3 @@
+namespace Frontend.Shared.Models.Dtos;
+
+public record PlayerDto(string Name, string ConnectionId, TimeSpan? TimeReserve = null) { }
