+++ Shared/Models/Dtos/UserDto.cs
@@ -1,0 +1,3 @@
+namespace Frontend.Shared.Models.Dtos;
+
+public record UserDto(string Id, string Name, int EloRating, string? Avatar = null, string AvatarUrl = "default-avatar.png") { }
