--- Shared/Models/PlayerDto.cs
@@ -1,8 +1,0 @@
-namespace Frontend.Shared.Models;
-
-public record PlayerDto
-{
-	public required string Name { get; set; }
-	public required string ConnectionId { get; set; }
-	public TimeSpan TimeReserve { get; set; }
-}
